@@ -13,16 +13,16 @@
   "private": true,
   "dependencies": {
     "@abacritt/angularx-social-login": "^1.2.1",
-    "@angular/animations": "~13.1.0",
+    "@angular/animations": "^14.0.6",
     "@angular/cdk": "13.2.6",
-    "@angular/common": "~13.1.0",
-    "@angular/compiler": "~13.1.0",
-    "@angular/core": "~13.1.0",
-    "@angular/forms": "~13.1.0",
+    "@angular/common": "^14.0.6",
+    "@angular/compiler": "^14.0.6",
+    "@angular/core": "^14.0.6",
+    "@angular/forms": "^14.0.6",
     "@angular/material": "13.2.6",
-    "@angular/platform-browser": "~13.1.0",
-    "@angular/platform-browser-dynamic": "~13.1.0",
-    "@angular/router": "~13.1.0",
+    "@angular/platform-browser": "^14.0.6",
+    "@angular/platform-browser-dynamic": "^14.0.6",
+    "@angular/router": "^14.0.6",
     "date-fns": "^2.28.0",
     "lodash-es": "^4.17.21",
     "rxjs": "~7.4.0",
@@ -30,9 +30,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.1.2",
-    "@angular/cli": "~13.1.2",
-    "@angular/compiler-cli": "~13.1.0",
+    "@angular-devkit/build-angular": "^14.0.6",
+    "@angular/cli": "^14.0.6",
+    "@angular/compiler-cli": "^14.0.6",
     "@types/jasmine": "~3.10.0",
     "@types/lodash-es": "^4.17.6",
     "@types/node": "^12.11.1",
@@ -42,6 +42,6 @@
     "karma-coverage": "~2.1.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.5.2"
+    "typescript": "~4.7.4"
   }
 }
